requirements.txt
python-dotenv
pymongo
sentence-transformers